+++ docs/my_iot/consts/sidebar-items.js
@@ -1,1 +1,1 @@
-initSidebarItems({"constant":[["ANDROID_CHROME_192",""],["ANDROID_CHROME_512",""],["APPLE_TOUCH_ICON",""],["FAVICON",""],["FAVICON_16",""],["FAVICON_32",""],["USER_AGENT",""]]});
+initSidebarItems({"constant":[["USER_AGENT",""]]});