+++ docs/my_iot/web/sidebar-items.js
@@ -1,1 +1,1 @@
-initSidebarItems({"fn":[["get_sensor","Get sensor page response."],["get_sensor_json","Get last sensor value JSON response."],["index","Get index page response."],["start_server","Start the web application."]]});
+initSidebarItems({"constant":[["ANDROID_CHROME_192",""],["ANDROID_CHROME_512",""],["APPLE_TOUCH_ICON",""],["FAVICON",""],["FAVICON_16",""],["FAVICON_32",""]],"fn":[["get_sensor","Get sensor page response."],["get_sensor_json","Get last sensor value JSON response."],["index","Get index page response."],["start_server","Start the web application."]]});